// scheduler-service/src/env.d.ts
import "dotenv/config";
export declare const ENV: {
  TEMPORAL_ADDRESS: string;
  TEMPORAL_NAMESPACE: string;
  TASK_QUEUE: string;
  TEMPORAL_API_KEY: string | undefined;
  REMINDER_SERVICE_BASEURL: string;
  USER_SERVICE_BASEURL: string;
  POLL_INTERVAL_MS: number;
  DUE_WINDOW_SEC: number;
  AUTH_BEARER: string;
  TZ: string;
};
//# sourceMappingURL=env.d.ts.map